rework the formatter_for, resource_for, paginator_for methods to use only safe_constantize.   This seems to work across all supported versions of Ruby.   Cleans up the code, and may fix issue #91
  Note: I do not remember the rational for the two separate versions
